+++ Runtime/tools/ITweenModule.cs
@@ -1,0 +1,52 @@
+using UnityEngine;
+
+public interface ITweenModule
+{
+    // //UI:
+    // static TweenerCore<float, float, FloatOptions> DOFade(this CanvasGroup target, float endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOColor(this Graphic target, Color endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOFade(this Graphic target, float endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOColor(this Image target, Color endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOFade(this Image target, float endValue, float duration);
+    // static TweenerCore<float, float, FloatOptions> DOFillAmount(this Image target, float endValue, float duration);
+    // static Sequence DOGradientColor(this Image target, Gradient gradient, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOFlexibleSize(this LayoutElement target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOMinSize(this LayoutElement target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOPreferredSize(this LayoutElement target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Color, Color, ColorOptions> DOColor(this Outline target, Color endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOFade(this Outline target, float endValue, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOScale(this Outline target, Vector2 endValue, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOAnchorPos(this RectTransform target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOAnchorPosX(this RectTransform target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOAnchorPosY(this RectTransform target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector3, Vector3, VectorOptions> DOAnchorPos3D(this RectTransform target, Vector3 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector3, Vector3, VectorOptions> DOAnchorPos3DX(this RectTransform target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector3, Vector3, VectorOptions> DOAnchorPos3DY(this RectTransform target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector3, Vector3, VectorOptions> DOAnchorPos3DZ(this RectTransform target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOAnchorMax(this RectTransform target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOAnchorMin(this RectTransform target, Vector2 endValue, float duration, bool snapping = false);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOPivot(this RectTransform target, Vector2 endValue, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOPivotX(this RectTransform target, float endValue, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOPivotY(this RectTransform target, float endValue, float duration);
+    // static TweenerCore<Vector2, Vector2, VectorOptions> DOSizeDelta(this RectTransform target, Vector2 endValue, float duration, bool snapping = false);
+    // static Tweener DOPunchAnchorPos(this RectTransform target, Vector2 punch, float duration, int vibrato = 10, float elasticity = 1, bool snapping = false);
+    // static Tweener DOShakeAnchorPos(this RectTransform target, float duration, float strength = 100, int vibrato = 10, float randomness = 90, bool snapping = false, bool fadeOut = true);
+    // static Tweener DOShakeAnchorPos(this RectTransform target, float duration, Vector2 strength, int vibrato = 10, float randomness = 90, bool snapping = false, bool fadeOut = true);
+    // static Sequence DOJumpAnchorPos(this RectTransform target, Vector2 endValue, float jumpPower, int numJumps, float duration, bool snapping = false);
+    // static Tweener DONormalizedPos(this ScrollRect target, Vector2 endValue, float duration, bool snapping = false);
+    // static Tweener DOHorizontalNormalizedPos(this ScrollRect target, float endValue, float duration, bool snapping = false);
+    // static Tweener DOVerticalNormalizedPos(this ScrollRect target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<float, float, FloatOptions> DOValue(this Slider target, float endValue, float duration, bool snapping = false);
+    // static TweenerCore<Color, Color, ColorOptions> DOColor(this Text target, Color endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOFade(this Text target, float endValue, float duration);
+    // static TweenerCore<string, string, StringOptions> DOText(this Text target, string endValue, float duration, bool richTextEnabled = true, ScrambleMode scrambleMode = ScrambleMode.None, string scrambleChars = null);
+    // static Tweener DOBlendableColor(this Graphic target, Color endValue, float duration);
+    // static Tweener DOBlendableColor(this Image target, Color endValue, float duration);
+    // static Tweener DOBlendableColor(this Text target, Color endValue, float duration);
+
+    // // Sprite:
+    // static TweenerCore<Color, Color, ColorOptions> DOColor(this SpriteRenderer target, Color endValue, float duration);
+    // static TweenerCore<Color, Color, ColorOptions> DOFade(this SpriteRenderer target, float endValue, float duration);
+    // static Sequence DOGradientColor(this SpriteRenderer target, Gradient gradient, float duration);
+    // static Tweener DOBlendableColor(this SpriteRenderer target, Color endValue, float duration);
+ }