+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 2 Oktober 2020:
 - Added a ReadOnly. This is a helpful functionality that makes a property 'ReadOnly' in the editor, so you can't accidently or purposely change it.
+
+22 February 2021:
+- Added a Visualizer
+- Added StringUtils
+- Added LowestAbs and GetPointOnCircle to SimpleMath.cs
+
+2 May 2021:
+- Added a EditorUtils.Divider() to draw a divider in the inspector